// src/core/new.swipe.engine/params.type.ts
export type DBSCANConfig = {
    epsilon: number
    minPoints: number
    weights: Record<string, never>
    distanceFunction: "euclidean" | "cosine" | "manhattan"
    randomSeed: number
    initMethod: "k-means++" | "random"
    threshold: number
}